Fixed a infinite redirection issue
When a user tried to access a page he/she dosn't have access to, he/she
is redirected to the login page. The login page was set to redirect the
user to next if already logged in. This resulted in an infinite
redirection loop. Now the login view does not automatically redirect the
user.